requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
breezyslam>=1.0.0
pyserial>=3.5
pandas>=1.3.0 